cloudconfig: move files
